4
Add visitors to the ride history using the addVisitorToHistory method. Use the checkVisitorFromHistory method to check if the visitor is playing history. Use the numberOfVisitors method to return the number of visitors in the ride history. Using the method PrintRideHistory, print the details of all the visitors who have ridden the ride, using an iterator. In "...
Experiments were carried out in the AssignmentTwo class. Removed some comments that had alerted me to the change.